feat: move flutter_finder under extensions
