+++ test/functional/flutter_integration/commands_test.py
@@ -16,7 +16,7 @@ import os
 
 from appium.webdriver.common.appiumby import AppiumBy
 from appium.webdriver.extensions.flutter_integration.scroll_directions import ScrollDirection
-from appium.webdriver.flutter_finder import FlutterFinder
+from appium.webdriver.extensions.flutter_integration.flutter_finder import FlutterFinder
 from test.functional.flutter_integration.helper.test_helper import BaseTestCase
