+++ test/functional/flutter_integration/finder_test.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from appium.webdriver.common.appiumby import AppiumBy
-from appium.webdriver.flutter_finder import FlutterFinder
+from appium.webdriver.extensions.flutter_integration.flutter_finder import FlutterFinder
 from test.functional.flutter_integration.helper.test_helper import BaseTestCase
 
 LOGIN_BUTTON_FINDER = FlutterFinder.by_text("Login")
