+++ test/unit/webdriver/flutter_integration/flutter_actions_test.py
@@ -19,7 +19,7 @@ import httpretty
 from appium.common.helper import encode_file_to_base64
 from appium.webdriver.extensions.flutter_integration.flutter_commands import FlutterCommand
 from appium.webdriver.extensions.flutter_integration.scroll_directions import ScrollDirection
-from appium.webdriver.flutter_finder import FlutterFinder
+from appium.webdriver.extensions.flutter_integration.flutter_finder import FlutterFinder
 from appium.webdriver.webelement import WebElement as MobileWebElement
 from test.unit.helper.test_helper import appium_command, flutter_w3c_driver, get_httpretty_request_body
